requirements_pip.txt
Django==1.7
MySQL-python==1.2.5
Werkzeug==0.9.6
argparse==1.2.1
distribute==0.6.24
geopy==1.4.0
python-firebase==1.2
requests==2.4.3
wsgiref==0.1.2
django-bootstrap3-datetimepicker==2.2.3

